Moved alert to components
