+++ routes/preview-api-route.tsx
@@ -3,9 +3,9 @@ import { useParams, type JSXElement } from "revolution";
 import { SitemapRoute } from "../plugins/sitemap.ts";
 import { useAppHtml } from "./app.html.tsx";
 import { Repository } from "../resources/repository.ts";
-import { Alert } from "./api-minor-index-route.tsx";
 import { ApiPage } from "./api-reference-route.tsx";
 import { createSibling } from "./links-resolvers.ts";
+import { Alert } from "../components/alert.tsx";
 
 export function previewApiRoute({
   library,
